Add: Login, register, JWT Auth

+++ routes/api.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\KendaraanController;
 use App\Http\Controllers\MotorController;
 use App\Http\Controllers\MobilController;
 use App\Http\Controllers\SalesController;
+use App\Http\Controllers\UserController;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,36 +23,39 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-
 Route::group(['prefix' => 'beta'], function () {
     // Kendaraan Routes
-    Route::get('/kendaraan', [KendaraanController::class, 'index']);
-    Route::post('/kendaraan', [KendaraanController::class, 'store']);
-    Route::put('/kendaraan/{id}', [KendaraanController::class, 'update']);
-    Route::get('/kendaraan/{id}', [KendaraanController::class, 'show']);
-    Route::delete('/kendaraan/{id}', [KendaraanController::class, 'destroy']);
+    Route::get('/kendaraan', [KendaraanController::class, 'index'])->middleware('jauth');
+    Route::post('/kendaraan', [KendaraanController::class, 'store'])->middleware('jauth');
+    Route::put('/kendaraan/{id}', [KendaraanController::class, 'update'])->middleware('jauth');
+    Route::get('/kendaraan/{id}', [KendaraanController::class, 'show'])->middleware('jauth');
+    Route::delete('/kendaraan/{id}', [KendaraanController::class, 'destroy'])->middleware('jauth');
 
     // Stock Routes
-    Route::get('/stock/kendaraan/all', [KendaraanController::class, 'stockAll']);
-    Route::get('/stock/kendaraan/{id}', [KendaraanController::class, 'stockKendaraan']);
+    Route::get('/stock/kendaraan/all', [KendaraanController::class, 'stockAll'])->middleware('jauth');
+    Route::get('/stock/kendaraan/{id}', [KendaraanController::class, 'stockKendaraan'])->middleware('jauth');
 
     // Motor Routes
-    Route::get('/motor', [MotorController::class, 'motorIndex']);
-    Route::post('/motor', [MotorController::class, 'motorStore']);
-    Route::put('/motor/{id}', [MotorController::class, 'motorUpdate']);
-    Route::get('/motor/{id}', [MotorController::class, 'motorShow']);
-    Route::delete('/motor/{id}', [MotorController::class, 'motorDestroy']);
+    Route::get('/motor', [MotorController::class, 'motorIndex'])->middleware('jauth');
+    Route::post('/motor', [MotorController::class, 'motorStore'])->middleware('jauth');
+    Route::put('/motor/{id}', [MotorController::class, 'motorUpdate'])->middleware('jauth');
+    Route::get('/motor/{id}', [MotorController::class, 'motorShow'])->middleware('jauth');
+    Route::delete('/motor/{id}', [MotorController::class, 'motorDestroy'])->middleware('jauth');
 
     // Mobil Routes
-    Route::get('/mobil', [MobilController::class, 'mobilIndex']);
-    Route::post('/mobil', [MobilController::class, 'mobilStore']);
-    Route::put('/mobil/{id}', [MobilController::class, 'mobilUpdate']);
-    Route::get('/mobil/{id}', [MobilController::class, 'mobilShow']);
-    Route::delete('/mobil/{id}', [MobilController::class, 'mobilDestroy']);
+    Route::get('/mobil', [MobilController::class, 'mobilIndex'])->middleware('jauth');
+    Route::post('/mobil', [MobilController::class, 'mobilStore'])->middleware('jauth');
+    Route::put('/mobil/{id}', [MobilController::class, 'mobilUpdate'])->middleware('jauth');
+    Route::get('/mobil/{id}', [MobilController::class, 'mobilShow'])->middleware('jauth');
+    Route::delete('/mobil/{id}', [MobilController::class, 'mobilDestroy'])->middleware('jauth');
 
     // Sales Route
-    Route::get('/sales', [SalesController::class, 'index']);
-    Route::get('/sales/kendaraan/{id}', [SalesController::class, 'showSalesKendaraan']);
-    Route::get('/sales/{id}', [SalesController::class, 'show']);
-    Route::post('/sales', [SalesController::class, 'store']);
+    Route::get('/sales', [SalesController::class, 'index'])->middleware('jauth');
+    Route::get('/sales/kendaraan/{id}', [SalesController::class, 'showSalesKendaraan'])->middleware('jauth');
+    Route::get('/sales/{id}', [SalesController::class, 'show'])->middleware('jauth');
+    Route::post('/sales', [SalesController::class, 'store'])->middleware('jauth');
+
+    // User Routes
+    Route::post('/registration', [UserController::class, 'store']);
+    Route::post('/login', [UserController::class, 'login']);
   });
